Fix bugs

+++ lib/galaxy/tool_util/verify/__init__.py
@@ -471,8 +471,8 @@ def get_image_metric(attributes):
     attributes = attributes or {}
     metrics = {
         "mae": lambda im1, im2: numpy.abs(im1 - im2).mean(),
-        "mse": lambda im1, im2: numpy.square(im1 - im2).mean(),
-        "rms": lambda im1, im2: math.sqrt(numpy.square(im1 - im2).mean()),
+        "mse": lambda im1, im2: numpy.square((im1 - im2).astype(float)).mean(),
+        "rms": lambda im1, im2: math.sqrt(numpy.square((im1 - im2).astype(float)).mean()),
         "fro": lambda im1, im2: numpy.linalg.norm(im1 - im2, "fro"),
         "iou": lambda im1, im2: 1 - intersection_over_union(im1, im2),
     }
